refactor(workspace): extract mirror ref probes into mirror/refs.rs
The Server Guards job runs scripts/check-file-size.sh with a 51200-byte
cap per Rust source file. mirror.rs was 51727 bytes — 527 over. It was
already at 47450 bytes (93% of the cap) on main, so adding branch_exists
and its four tests was simply the change that pushed it past.

The guard exists to force decomposition, so this splits the file rather
than trimming it. branch_ahead_of_base and branch_exists are both
read-only probes of refs/heads/* in the bare mirror: same shape, no
mutation, and a single test module already covered exactly the two of
them. That is a cohesive unit distinct from mirror lifecycle
(create/fetch/gc/clone), which is what the rest of mirror.rs does.

Both are inherent `impl MirrorManager` methods, so moving them to a
child module changes no paths and no call sites — the public API is
byte-for-byte identical and lib.rs needs no new re-export.

All reasoning comments move verbatim, including the note on why an
absent branch cannot be classified off the MirrorError variant
(MirrorError::Missing means the mirror directory is absent, while an
absent branch surfaces as MirrorError::Git, same as a transient
failure). Losing that would re-open a defect affecting every
first-cycle run.

Renamed the moved test module ahead_of_base_tests -> tests: it now
covers branch_exists too, and it is the only test module in refs.rs.

  mirror.rs       51727 -> 38736 bytes (1324 -> 1005 lines)
  mirror/refs.rs  new, 13819 bytes (338 lines)

Guard passes over all 1203 in-scope files. Test counts are unchanged
from the base commit (329 run, 275 passed, 54 failed — all 54 are
read_source DB tests that need a Postgres test DB, failing identically
before and after this change).
